Move logic for escaping command arguments to own class
