+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/Debug/ProjectLaunchTargetsProvider.cs
@@ -30,7 +30,6 @@ internal class ProjectLaunchTargetsProvider :
     IDebugProfileLaunchTargetsProvider3,
     IDebugProfileLaunchTargetsProvider4
 {
-    private static readonly char[] s_escapedChars = new[] { '^', '<', '>', '&' };
     private readonly ConfiguredProject _project;
     private readonly IUnconfiguredProjectVsServices _unconfiguredProjectVsServices;
     private readonly IDebugTokenReplacer _tokenReplacer;
@@ -193,9 +192,11 @@ internal class ProjectLaunchTargetsProvider :
     {
         if (useCmdShell)
         {
-            // Escape the characters ^<>& so that they are passed to the application rather than interpreted by cmd.exe.
-            string? escapedArgs = EscapeString(debugArgs, s_escapedChars);
-            finalArguments = $"/c \"\"{debugExe}\" {escapedArgs} & pause\"";
+            // Any debug arguments must be escaped.
+            finalArguments = debugArgs is null
+                ? $"/c \"\"{debugExe}\" & pause\""
+                : $"/c \"\"{debugExe}\" {CommandEscaping.EscapeString(debugArgs)} & pause\"";
+
             finalExePath = Path.Combine(Environment.SystemDirectory, "cmd.exe");
         }
         else
@@ -461,93 +462,6 @@ internal class ProjectLaunchTargetsProvider :
         string framework = await properties.GetEvaluatedPropertyValueAsync(ConfigurationGeneral.TargetFrameworkIdentifierProperty);
 
         return GetManagedDebugEngineForFramework(framework);
-    }
-
-    /// <summary>
-    /// Escapes the given characters in a given string, ignoring escape sequences when inside a quoted string.
-    /// </summary>
-    /// <param name="unescaped">The string to escape.</param>
-    /// <param name="toEscape">The characters to escape in the string.</param>
-    /// <returns>The escaped string.</returns>
-    [return: NotNullIfNotNull(nameof(unescaped))]
-    internal static string? EscapeString(string? unescaped, char[] toEscape)
-    {
-        if (Strings.IsNullOrWhiteSpace(unescaped))
-        {
-            return unescaped;
-        }
-
-        bool ShouldEscape(char c)
-        {
-            foreach (char escapeChar in toEscape)
-            {
-                if (escapeChar == c)
-                    return true;
-            }
-            return false;
-        }
-
-        StringState currentState = StringState.NormalCharacter;
-        var finalBuilder = PooledStringBuilder.GetInstance();
-        foreach (char currentChar in unescaped)
-        {
-            switch (currentState)
-            {
-                case StringState.NormalCharacter:
-                    // If we're currently not in a quoted string, then we need to escape anything in toEscape.
-                    // The valid transitions are to EscapedCharacter (for a '\', such as '\"'), and QuotedString.
-                    if (currentChar == '\\')
-                    {
-                        currentState = StringState.EscapedCharacter;
-                    }
-                    else if (currentChar == '"')
-                    {
-                        currentState = StringState.QuotedString;
-                    }
-                    else if (ShouldEscape(currentChar))
-                    {
-                        finalBuilder.Append('^');
-                    }
-
-                    break;
-                case StringState.EscapedCharacter:
-                    // If a '\' was the previous character, then we blindly append to the string, escaping if necessary,
-                    // and move back to NormalCharacter. This handles '\"'
-                    if (ShouldEscape(currentChar))
-                    {
-                        finalBuilder.Append('^');
-                    }
-
-                    currentState = StringState.NormalCharacter;
-                    break;
-                case StringState.QuotedString:
-                    // If we're in a string, we don't escape any characters. If the current character is a '\',
-                    // then we move to QuotedStringEscapedCharacter. This handles '\"'. If the current character
-                    // is a '"', then we're out of the string. Otherwise, we stay in the string.
-                    if (currentChar == '\\')
-                    {
-                        currentState = StringState.QuotedStringEscapedCharacter;
-                    }
-                    else if (currentChar == '"')
-                    {
-                        currentState = StringState.NormalCharacter;
-                    }
-
-                    break;
-                case StringState.QuotedStringEscapedCharacter:
-                    // If we have one slash, then we blindly append to the string, no escaping, and move back to
-                    // QuotedString. This handles escaped '"' inside strings.
-                    currentState = StringState.QuotedString;
-                    break;
-                default:
-                    // We can't get here.
-                    throw new InvalidOperationException();
-            }
-
-            finalBuilder.Append(currentChar);
-        }
-
-        return finalBuilder.ToStringAndFree();
     }
 
     /// <summary>
